Implement the BOM removal and rails rake task hook
We look at each css file in the rails asset location and if we detect a
BOM we remove it from the file and also regenerate the gzipped version
(if it exists) using the new BOM-less version of the CSS file (using the
same bit of sprockets that rails uses to do so).

It leaves files without a BOM file alone, and leaves files that are not
CSS files alone too.

We hook into rails by enhancing the `assets:precompile` task to invoke
ours once it has completed.